Show last unlocked level on every page
Bug:
The menu only shows the next unlocked level if you are on the last
completed level's page.
e.g. If you have unlocked Level5 and are on page Level1, you will not
see Level5 in the menu.

Fix:
I modified the menu a little to show the last unlocked level no matter
which page you are on.

Tested:
- New session on Tutorial: shows nothing (same as right now)
- Level1 completed, on page Tutorial: Shows up to Level2 (would
previously show up to Level1)
- Level1 completed, on page Level1: Shows up to Level2 (same as right
now)
- Level1 completed, on page Level2: Shows up to Level2 (would previously
show up to Level1)
- Level1 completed, on page Level20: Shows up to Level2 (would
previously show up to Level1)

Fixes #295