Remove locked tools from Level1 and Level2
Level one is to construct an equilateral triangle and unlock that tool,
but that tool is immediately available at the beginning of the lesson.

On top of that, using it directly will show "Completed" (technically
correct), but not register the level as finished (or show the link to
Level2).

Also a similar fix in Level2 for Midpoint (this one registers it as a
completed level in 1 move..!).

Fixes 354